Move Part class to type package
