+++ firebase-ai-kmp/src/androidMain/kotlin/io/github/seanchinjunkai/firebase/ai/type/PlatformImage.android.kt
@@ -1,4 +1,4 @@
-package io.github.seanchinjunkai.firebase.ai
+package io.github.seanchinjunkai.firebase.ai.type
 
 import android.graphics.Bitmap
