--- firebase-ai-kmp/src/commonMain/kotlin/io/github/seanchinjunkai/firebase/ai/PlatformImage.kt
@@ -1,3 +1,0 @@
-package io.github.seanchinjunkai.firebase.ai
-
-expect class PlatformImage
