+++ firebase-ai-kmp/src/commonMain/kotlin/io/github/seanchinjunkai/firebase/ai/type/Part.kt
@@ -1,4 +1,6 @@
-package io.github.seanchinjunkai.firebase.ai
+package io.github.seanchinjunkai.firebase.ai.type
+
+import io.github.seanchinjunkai.firebase.ai.type.PlatformImage
 
 public interface Part {}
