+++ firebase-ai-kmp/src/commonMain/kotlin/io/github/seanchinjunkai/firebase/ai/type/PlatformImage.kt
@@ -1,0 +1,3 @@
+package io.github.seanchinjunkai.firebase.ai.type
+
+expect class PlatformImage
