+++ firebase-ai-kmp/src/iosMain/kotlin/io/github/seanchinjunkai/firebase/ai/type/PlatformImage.ios.kt
@@ -1,4 +1,4 @@
-package io.github.seanchinjunkai.firebase.ai
+package io.github.seanchinjunkai.firebase.ai.type
 
 import platform.UIKit.UIImage
